refactor(pt-94): extract ScrollManager into a separate component
